Optimiser works; reproduces Fig 3!
Key thing, as well as fixing the sinh() numerics, was to make us of (7)
and express a2 in terms of Omega.
Thereby the optimisation becomes a simple 1D one, and it's not trying to
escape the bounds. I guess there's some kind of singularity which is
messing up the direct numerics.